+++ packages/docs/src/stories/components/combobox.stories.ts
@@ -221,6 +221,29 @@ export const Disabled = {
 };
 
 /**
+ * Use the `visually-disabled` attribute to style the component as if it was disabled and enable `aria-disabled` to allow it to be reachable by screen readers.
+ *
+ * __Hint:__ When using this attribute, make sure to provide ways to inform the user why the element is disabled and how to enable it. This can be done by using the `help-text` attribute or wrapping the element in a sd-tooltip.
+ *
+ * **Accessibility Hint:** Disabling elements is not recommended for accessibility reasons.
+ */
+export const VisuallyDisabled = {
+  render: () => html`
+    <div class="w-[400px] h-[200px] pt-12">
+      <sd-tooltip content="Visually Disabled" trigger="click focus" size="sm" placement="top">
+        <sd-combobox size="lg" label="Label" placeholder="Disabled" placement="bottom" value="" visually-disabled>
+          <sd-option value="option-1">Option 1</sd-option>
+          <sd-option value="option-2">Option 2</sd-option>
+          <sd-option value="option-3">Option 3</sd-option>
+          <sd-option value="option-4">Option 4</sd-option>
+          <sd-option value="option-5">Option 5</sd-option>
+        </sd-combobox>
+      </sd-tooltip>
+    </div>
+  `
+};
+
+/**
  * Use the `help-text` attribute to provide additional context or instructions.
  *
  * For help texts that contain HTML, use the `help-text` slot instead.