+++ packages/docs/src/stories/components/link.stories.ts
@@ -145,6 +145,30 @@ export const Disabled = {
 };
 
 /**
+ * Use the `visually-disabled` attribute to style the component as if it was disabled and enable `aria-disabled` to allow it to be reachable by screen readers.
+ *
+ * __Hint:__ When using this attribute, make sure to provide ways to inform the user why the element is disabled and how to enable it. This can be done by using the `help-text` attribute or wrapping the element in a sd-tooltip.
+ *
+ * **Accessibility Hint:** Disabling elements is not recommended for accessibility reasons.
+ */
+export const VisuallyDisabled = {
+  name: 'Visually Disabled',
+  render: () => html`
+    <div class="flex gap-12 h-[100px] pt-12">
+      <sd-tooltip content="Visually Disabled" trigger="click focus" size="sm">
+        <sd-link
+          size="lg"
+          href="https://solid-design-system.fe.union-investment.de/docs/"
+          visually-disabled
+          class="visually-disabled-link"
+          >Visually Disabled</sd-link
+        >
+      </sd-tooltip>
+    </div>
+  `
+};
+
+/**
  * Use the `inverted` attribute when displayed on primary background.
  */
 export const Inverted = {